add resources requests and limits to nodePoolOptions
