+++ pkg/config.go
@@ -2,91 +2,12 @@ package v1
 
 import (
 	"encoding/base64"
-	"fmt"
 	"github.com/onepanelio/core/pkg/util"
-	"github.com/onepanelio/core/pkg/util/ptr"
 	log "github.com/sirupsen/logrus"
 	"google.golang.org/grpc/codes"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/yaml"
-	"strings"
 )
-
-// SystemConfig is configuration loaded from kubernetes config and secrets that includes information about the
-// database, server, etc.
-type SystemConfig map[string]string
-
-// GetValue returns the value in the underlying map if it exists, otherwise nil is returned
-// If the value does not exist, it is also logged.
-func (s SystemConfig) GetValue(name string) *string {
-	value, ok := s[name]
-	if !ok {
-		log.WithFields(log.Fields{
-			"Method": "SystemConfig.GetValue",
-			"Name":   name,
-			"Error":  "does not exist",
-		})
-
-		return nil
-	}
-
-	return &value
-}
-
-// Domain gets the ONEPANEL_DOMAIN value, or nil.
-func (s SystemConfig) Domain() *string {
-	return s.GetValue("ONEPANEL_DOMAIN")
-}
-
-// APIURL gets the ONEPANEL_API_URL, or nil.
-func (s SystemConfig) APIURL() *string {
-	return s.GetValue("ONEPANEL_API_URL")
-}
-
-// APIProtocol returns either http:// or https:// or nil.
-// It is based on the ONEPANEL_API_URL config value and checks if it has https or http
-func (s SystemConfig) APIProtocol() *string {
-	url := s.APIURL()
-	if url == nil {
-		return nil
-	}
-
-	if strings.HasPrefix(*url, "https://") {
-		return ptr.String("https://")
-	}
-
-	return ptr.String("http://")
-}
-
-// FQDN gets the ONEPANEL_FQDN value or nil.
-func (s SystemConfig) FQDN() *string {
-	return s.GetValue("ONEPANEL_FQDN")
-}
-
-// NodePoolLabel gets the applicationNodePoolLabel from the config or returns nil.
-func (s SystemConfig) NodePoolLabel() (label *string) {
-	return s.GetValue("applicationNodePoolLabel")
-}
-
-// NodePoolOptions loads and parses the applicationNodePoolOptions from the config.
-// If there is no data, an error is returned.
-func (s SystemConfig) NodePoolOptions() (options []*ParameterOption, err error) {
-	data := s.GetValue("applicationNodePoolOptions")
-	if data == nil {
-		return nil, fmt.Errorf("no nodePoolOptions in config")
-	}
-
-	if err = yaml.Unmarshal([]byte(*data), &options); err != nil {
-		return
-	}
-
-	return
-}
-
-// DatabaseDriverName gets the databaseDriverName value, or nil.
-func (s SystemConfig) DatabaseDriverName() *string {
-	return s.GetValue("databaseDriverName")
-}
 
 func (c *Client) getConfigMap(namespace, name string) (configMap *ConfigMap, err error) {
 	cm, err := c.CoreV1().ConfigMaps(namespace).Get(name, metav1.GetOptions{})
